% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_single_field_session_counts.R
\name{get_single_field_session_counts}
\alias{get_single_field_session_counts}
\title{Get Distinct Field Session Counts}
\usage{
get_single_field_session_counts(data, field)
}
\arguments{
\item{data}{This is your full dataset that you are looking to compare.}

\item{field}{This is the column name that you are looking to get unique session counts for.}
}
\description{
This counts distinct sessionId values for any dimension in your query. You must have your sessions in a column named sessionId for this to work. The output will be the values, the counts, and then percentage.
}
