% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent_change.R
\name{percent_change}
\alias{percent_change}
\title{Percent Change}
\usage{
percent_change(old, new)
}
\arguments{
\item{old}{The older value being compared.}

\item{new}{The newer value being compared.}
}
\description{
This compares change between two values.
}
\examples{
percent_change()
}
